src/requirements.txt
ipython==8.4.0
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.5
opencv_python==4.5.5.64
Pillow==9.2.0
scikit_image==0.19.2
scipy==1.7.3
skimage==0.0
torch==1.11.0
torchvision==0.12.0
